+++ requirements.lock
@@ -3,3 +3,4 @@ wtforms==2.2.1
 passlib==1.7.2
 flask_mysqldb==0.2.0
 mailchimp3==3.0.12
+requests==2.22.0
