+++ requirements.txt
@@ -3,3 +3,4 @@ wtforms
 passlib
 flask_mysqldb
 mailchimp3
+requests
